requirements.txt
SpeechRecognition
pyttsx3
requests
wikipedia
python-dotenv
